Fix kubeadmin logout redirect
