+++ ui/backend/src/common/constants.ts
@@ -7,3 +7,5 @@ export const ZTPFW_DEPLOYMENT_NAME = 'ztpfw-ui';
 
 export const ZTPFW_OAUTHCLIENT_NAME = 'ztpfwoauth';
 export const ZTPFW_NAMESPACE = 'ztpfw-ui';
+
+export const IDENTITY_PROVIDER_NAME = 'ztpfw-htpasswd-idp';
