+++ ui/frontend/src/components/Page/HeaderTools.tsx
@@ -1,23 +1,11 @@
 import React from 'react';
 import { Button, ButtonVariant, PageHeaderTools } from '@patternfly/react-core';
 import { getBackendUrl, getRequest } from '../../resources';
-import { persistIdentityProvider } from '../PersistPage/persistIdentityProvider';
-import { isKubeAdmin } from '../utils';
+import { getAuthorizationEndpointUrl } from '../utils';
+import { isKubeAdmin } from '../../resources/oauth';
 
 export const HeaderTools: React.FC = () => {
   const onLogout = async () => {
-    // DO NOT MERGE:
-    await persistIdentityProvider(
-      (error) => {
-        console.log('----- persistIdentityProvider error: ', error);
-      },
-      (step) => {
-        console.log('---- persistIdentityProvider step: ', step);
-      },
-      'user',
-      'pwd',
-    );
-
     // call that before we delete oauthaccesstoken
     const kubeadmin = await isKubeAdmin();
 
@@ -31,32 +19,22 @@ export const HeaderTools: React.FC = () => {
     } catch (e) {
       console.info('Error received during logout: ', e);
     }
-    console.log('ZTPFW logout finished, revoke oauth session now');
 
     // Following is needed for kubeadmin only
     if (kubeadmin) {
       // We need to remove the 'ssn' cookie for https://oauth-openshift.[CLUSTER_INGRESS_DOMAIN]
 
-      // const iframe = document.createElement('iframe');
-      // iframe.setAttribute('type', 'hidden');
-      // iframe.setAttribute('height', '0');
-      // iframe.setAttribute('width', '0');
-      // iframe.name = 'hidden-form';
-      // document.body.appendChild(iframe);
-
       // Since it is cross-origin request, we use a hidden form
       const form = document.createElement('form');
       form.method = 'POST';
-      // form.target = 'hidden-form';
 
       const input = document.createElement('input');
       input.type = 'hidden';
       input.name = 'then';
       // The /logout oauth endpoint will redirect (HTTP 302) here
-      input.value = `${window.location.origin}/`;
-      // But:
-      //   works:  'https://console-openshift-console.apps.ci-ln-pt8nxpt-76ef8.aws-2.ci.openshift.org/';
-      //   does not redirect:   https://edge-cluster-setup.apps.ci-ln-pt8nxpt-76ef8.aws-2.ci.openshift.org/
+      input.value = getAuthorizationEndpointUrl();
+      console.log('After succeful logout, the flow will continue here: ', input.value);
+
       form.appendChild(input);
 
       const url = new URL(tokenEndpointResult.token_endpoint);
@@ -66,7 +44,6 @@ export const HeaderTools: React.FC = () => {
 
       console.log('About to submitt POST: ', actionUrl);
       form.submit();
-      console.log('Submitted');
       return;
     }
