+++ ui/frontend/src/components/PersistPage/constants.ts
@@ -10,8 +10,6 @@ export const UI_POD_NOT_READY = 'Configuration UI pod is not ready';
 export const API_LIVENESS_FAILED_TITLE = 'API can not be reached';
 export const WAIT_ON_OPERATOR_TITLE = 'Reading operator status failed';
 
-export const IDENTITY_PROVIDER_NAME = 'ztpfw-htpasswd-idp';
-
 export const DELAY_BEFORE_FINAL_REDIRECT = 10 * 1000;
 export const DELAY_BEFORE_QUERY_RETRY = 5 * 1000; /* ms */
 export const MAX_LIVENESS_CHECK_COUNT = 20 * ((60 * 1000) / DELAY_BEFORE_QUERY_RETRY); // max 20 minutes
